requirements.txt
boto3==1.29.5
botocore==1.32.5
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7
